+++ scripts/dom/labels.js
@@ -6,6 +6,32 @@ function masterFolder(folder, index)
 	dom.loadIndexPage(true);
 }
 
+function setFavorite(path)
+{
+	let favorites = storage.get('favorites');
+
+	if(favorites[path])
+		delete favorites[path];
+	else
+		favorites[path] = {added: time()};
+
+	storage.set('favorites', favorites);
+
+	let prevIndexLabel = dom.prevIndexLabel();
+
+	if(prevIndexLabel.favorites)
+		dom.reloadIndex();
+}
+
+function favorites()
+{
+	dom.selectMenuItem('favorites');
+	dom.setIndexLabel({favorites: true});
+	dom.loadIndexPage(true);
+}
+
 module.exports = {
 	masterFolder: masterFolder,
+	setFavorite: setFavorite,
+	favorites: favorites,
 };